Integrate single step tests for the CPU and fix the bugs found.
This uses the tests from
https://github.com/SingleStepTests/65x02/tree/main/nes6502. There are
10,000 for each opcode. We have implemented 172 opcodes so it runs a
total of around 1.7 million test cases. None of the bugs uncovered have
been fixed in this commit. The test cases take up almost 600 MB after
minification so they have not been added to the repo.

Note that we compile tests with opt-level=2. This is needed for serde to
deserialize 600 MB of JSON in a reasonable amount of time. Perhaps we
could look into some zero-copy serialization framework for this, but it
seems like too much work.

Fixes:

- Make sure bit 5 is stored as 1 internally.

This fixes the failing single step tests for PLP.

- Opcode 0x01 should use indexed indirect, not indirect indexed.

This was me misreading the opcode matrix. Caught by the single step
tests.

- Add more nop opcodes.

- Set negative flags correctly in various instructions.

This was a copy-paste error. Caught by the single step tests.

- Store the result in ASL, not temp.

Very funny mistake. Caught by single step tests.

- Set carry bit after calculating result in ROL.

Setting it before was incorrect and was caught by the single step tests.
Also, move setting the carry bit down in other instructions for
consistency.

- Keep Status impl in cpu.rs and other reorganization.

- Test emulated cycle count too and fix 2 opcodes.

The 2 opcodes were 0x28 and 0x08. Both returned a cycle count of 2 when
the actual cycle count should have been 4.