ppu: complete background rendering.
This required changes to the entire architecture of the emulator. Now,
NesBus owns a Ppu and a Cartridge but Ppu does not depend on NesBus at
all. It takes mutable refs to a Cartridge when necessary. This means the
PpuBus trait can be deleted. This also allows the API surface of the
library to be reduced. Also, I had to change the frontend to work with
this new architecture.

I added a function to util to extract a bit from u8.

I also implemented init and NMI handling in the CPU.

Adding PPU reading and writing to the bus required Bus::read to become a
function that takes &mut self. This is because reading PPU registers can
mutate the PPU's state. There are a lot of function signatures that were
updated following this change.